Move endpoints and api_key to config/setup
