+++ pkg/config/setup/endpoints_test.go
@@ -3,13 +3,12 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2016-present Datadog, Inc.
 
-package utils
+package setup
 
 import (
 	"testing"
 
 	"github.com/DataDog/datadog-agent/comp/core/secrets"
-	"github.com/DataDog/datadog-agent/pkg/config"
 	"github.com/DataDog/datadog-agent/pkg/util/optional"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
@@ -20,10 +19,10 @@ import (
 // endpoints are configured.
 // Refer to https://github.com/DataDog/viper/pull/2 for more details.
 func TestSecretBackendWithMultipleEndpoints(t *testing.T) {
-	conf := config.SetupConf()
+	conf := Conf()
 	conf.SetConfigFile("./tests/datadog_secrets.yaml")
 	// load the configuration
-	_, err := config.LoadDatadogCustom(conf, "datadog_secrets.yaml", optional.NewNoneOption[secrets.Component](), nil)
+	_, err := LoadDatadogCustom(conf, "datadog_secrets.yaml", optional.NewNoneOption[secrets.Component](), nil)
 	assert.NoError(t, err)
 
 	expectedKeysPerDomain := map[string][]string{
@@ -46,7 +45,7 @@ additional_endpoints:
   - someapikey
 `
 
-	testConfig := config.SetupConfFromYAML(datadogYaml)
+	testConfig := ConfFromYAML(datadogYaml)
 
 	multipleEndpoints, err := GetMultipleEndpoints(testConfig)
 
@@ -78,7 +77,7 @@ additional_endpoints:
   - someapikey
 `
 
-	testConfig := config.SetupConfFromYAML(datadogYaml)
+	testConfig := ConfFromYAML(datadogYaml)
 
 	multipleEndpoints, err := GetMultipleEndpoints(testConfig)
 
@@ -101,7 +100,7 @@ func TestGetMultipleEndpointsEnvVar(t *testing.T) {
 	t.Setenv("DD_API_KEY", "fakeapikey")
 	t.Setenv("DD_ADDITIONAL_ENDPOINTS", "{\"https://foo.datadoghq.com\": [\"someapikey\"]}")
 
-	testConfig := config.SetupConf()
+	testConfig := Conf()
 
 	multipleEndpoints, err := GetMultipleEndpoints(testConfig)
 
@@ -131,7 +130,7 @@ additional_endpoints:
   - someapikey
 `
 
-	testConfig := config.SetupConfFromYAML(datadogYaml)
+	testConfig := ConfFromYAML(datadogYaml)
 
 	multipleEndpoints, err := GetMultipleEndpoints(testConfig)
 
@@ -158,7 +157,7 @@ dd_url: "https://app.datadoghq.com"
 api_key: fakeapikey
 `
 
-	testConfig := config.SetupConfFromYAML(datadogYaml)
+	testConfig := ConfFromYAML(datadogYaml)
 
 	multipleEndpoints, err := GetMultipleEndpoints(testConfig)
 
@@ -186,7 +185,7 @@ additional_endpoints:
   - ""
 `
 
-	testConfig := config.SetupConfFromYAML(datadogYaml)
+	testConfig := ConfFromYAML(datadogYaml)
 
 	multipleEndpoints, err := GetMultipleEndpoints(testConfig)
 
@@ -219,7 +218,7 @@ additional_endpoints:
   - someapikey
 `
 
-	testConfig := config.SetupConfFromYAML(datadogYaml)
+	testConfig := ConfFromYAML(datadogYaml)
 
 	multipleEndpoints, err := GetMultipleEndpoints(testConfig)
 
@@ -241,7 +240,7 @@ additional_endpoints:
 func TestSiteEnvVar(t *testing.T) {
 	t.Setenv("DD_API_KEY", "fakeapikey")
 	t.Setenv("DD_SITE", "datadoghq.eu")
-	testConfig := config.SetupConfFromYAML("")
+	testConfig := ConfFromYAML("")
 
 	multipleEndpoints, err := GetMultipleEndpoints(testConfig)
 	externalAgentURL := GetMainEndpoint(testConfig, "https://external-agent.", "external_config.external_agent_dd_url")
@@ -261,7 +260,7 @@ func TestDefaultSite(t *testing.T) {
 	datadogYaml := `
 api_key: fakeapikey
 `
-	testConfig := config.SetupConfFromYAML(datadogYaml)
+	testConfig := ConfFromYAML(datadogYaml)
 
 	multipleEndpoints, err := GetMultipleEndpoints(testConfig)
 	externalAgentURL := GetMainEndpoint(testConfig, "https://external-agent.", "external_config.external_agent_dd_url")
@@ -282,7 +281,7 @@ func TestSite(t *testing.T) {
 site: datadoghq.eu
 api_key: fakeapikey
 `
-	testConfig := config.SetupConfFromYAML(datadogYaml)
+	testConfig := ConfFromYAML(datadogYaml)
 
 	multipleEndpoints, err := GetMultipleEndpoints(testConfig)
 	externalAgentURL := GetMainEndpoint(testConfig, "https://external-agent.", "external_config.external_agent_dd_url")
@@ -302,7 +301,7 @@ func TestDDURLEnvVar(t *testing.T) {
 	t.Setenv("DD_API_KEY", "fakeapikey")
 	t.Setenv("DD_URL", "https://app.datadoghq.eu")
 	t.Setenv("DD_EXTERNAL_CONFIG_EXTERNAL_AGENT_DD_URL", "https://custom.external-agent.datadoghq.com")
-	testConfig := config.SetupConfFromYAML("")
+	testConfig := ConfFromYAML("")
 	testConfig.BindEnv("external_config.external_agent_dd_url")
 
 	multipleEndpoints, err := GetMultipleEndpoints(testConfig)
@@ -323,7 +322,7 @@ func TestDDDDURLEnvVar(t *testing.T) {
 	t.Setenv("DD_API_KEY", "fakeapikey")
 	t.Setenv("DD_DD_URL", "https://app.datadoghq.eu")
 	t.Setenv("DD_EXTERNAL_CONFIG_EXTERNAL_AGENT_DD_URL", "https://custom.external-agent.datadoghq.com")
-	testConfig := config.SetupConfFromYAML("")
+	testConfig := ConfFromYAML("")
 	testConfig.BindEnv("external_config.external_agent_dd_url")
 
 	multipleEndpoints, err := GetMultipleEndpoints(testConfig)
@@ -348,7 +347,7 @@ func TestDDURLAndDDDDURLEnvVar(t *testing.T) {
 	t.Setenv("DD_URL", "https://app.datadoghq.dd_url.eu")
 
 	t.Setenv("DD_EXTERNAL_CONFIG_EXTERNAL_AGENT_DD_URL", "https://custom.external-agent.datadoghq.com")
-	testConfig := config.SetupConfFromYAML("")
+	testConfig := ConfFromYAML("")
 	testConfig.BindEnv("external_config.external_agent_dd_url")
 
 	multipleEndpoints, err := GetMultipleEndpoints(testConfig)
@@ -374,7 +373,7 @@ api_key: fakeapikey
 external_config:
   external_agent_dd_url: "https://external-agent.datadoghq.com"
 `
-	testConfig := config.SetupConfFromYAML(datadogYaml)
+	testConfig := ConfFromYAML(datadogYaml)
 
 	multipleEndpoints, err := GetMultipleEndpoints(testConfig)
 	externalAgentURL := GetMainEndpoint(testConfig, "https://external-agent.", "external_config.external_agent_dd_url")
@@ -398,7 +397,7 @@ api_key: fakeapikey
 external_config:
   external_agent_dd_url: "https://custom.external-agent.datadoghq.eu"
 `
-	testConfig := config.SetupConfFromYAML(datadogYaml)
+	testConfig := ConfFromYAML(datadogYaml)
 
 	multipleEndpoints, err := GetMultipleEndpoints(testConfig)
 	externalAgentURL := GetMainEndpoint(testConfig, "https://external-agent.", "external_config.external_agent_dd_url")
